Ajustes

+++ BuscaEmprego/BuscaEmprego.Entities/Endereco.cs
@@ -30,7 +30,7 @@ namespace BuscaEmprego.Entities
         [StringLength(255)]
         public string Tipo { get; set; }
 
-        public int Telefone { get; set; }
+        public long Telefone { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Usuario> Usuario { get; set; }
